Introduce refcount functions
The function CRYPTO_add{,_lock}() is used as a reference-counting
mechanism via locking of the whole subsystem. We can do this much better
by asking the compiler to use an intrinsic to atomically increment or
decrement the reference count.

As a way to break as little as possible in one go, introduce two new
functions which describe their goal which use these intrinsic. We will
then convert the call sites from CRYPTO_add() to this function.