Don't lock on the session when incrementing the refcount
The current comment says that it performs a full lock so the session
doesn't get freed between reading the pointer and incrementing the
refcount. Such a situation would however indicate bad reference
counting, or some code freeing a session to which it doesn't hold a
reference. This would be a bug in the caller and should be fixed there.