err: replace some locks with atomics
The CRYPTO_LOCK_ERR lock is often used to perform a compare-and-swap or
atomic exchange. Replace the locking in those places with the atomic
intrinsics which perform the same task.

This is still some of the more trivial replacements. There are a few
places where the critical section needs the global data structures to
stay still for a while which will require more work.