Replace locking callbacks via atomic operations
Instead of using a lock to atomically add or subtract from a number in
CRYPTO_add_lock(), replace the function body with an intrinsic which
atomically increments or decrements the value.

Instead of using a user-provided callback for the locks, implement the
coarse locks as spinlocks, which makes the library thread-safe by
default.