Trust the getaddrinfo error
If the call to getaddrinfo fails, we go to the code which uses
BIO_get_port which uses getaddrinfo again, so we haven't won anything.

The next bit of code calls BIO_get_host_ip which calls gethostbyanme,
which is in essence the same but not thread-safe. We haven't won
anything either.

Just remove the code, if getaddrinfo fails, then we won't connect.