Use getaddrinfo directory in conn_state
Instead of splitting looking up the address info into two different
states and playing around with a buffer for the IP and duplicating the
sockaddr_in structure, use getaddrinfo and fill the data structure which
is made for the purpose.

This makes the GET_IP and GET_PORT states the same, put them together in
the BEFORE state.